requirements.txt
gradio>=4.0.0
pymongo
python-dotenv
openai
requests
Pillow